+++ NEWS.md
@@ -1,5 +1,7 @@
 # paletteer 0.1.0.9000
 
+* Description of how to add palettes have been moved from a vignette to CONTRIBUTION.md.
+
 # paletteer 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.