+++ NEWS.md
@@ -1,4 +1,4 @@
-# paletteer 0.1.0.9000
+# paletteer 0.2.0.
 
 * Description of how to add palettes have been moved from a vignette to CONTRIBUTION.md.
 * 15 palettes have been added from the harrypotter package.